Prepare for Rain Deployment on Polygon (#31)
* Rain LoC Simulation (#23) (#29)

* feat: create Polygon simulation that deploys and configures a line of credit.

* feat: simulation handles manual deployment, full repayment cycle, and transfer of ownership back.

* feat: different deploy scripts for different chains

* fix: adjust polygon and mainnet simulations with new oracle and converting to push instead of pull payments.

---------


* Upgrade solc-version in foundry.toml to 0.8.20 (#30)

* Rain LoC Simulation (#23)

* feat: create Polygon simulation that deploys and configures a line of credit.

* feat: simulation handles manual deployment, full repayment cycle, and transfer of ownership back.

* feat: different deploy scripts for different chains

* fix: update the mainnet simulation.

* fix: adjust polygon simulation with new oracle and converting to push instead of pull payments.

* fix: modify the mainnet simulation for Rain to use push payment instead of a pull payment.

---------


* feat: upgrade solc-version to 0.8.20.

* fix: change IOracle.sol back to recently deployed version.

---------


* deployed factory and test Rain LoC

* mock Rain contracts

* added Mock to contract title

* added increaseNonce to Rain collateral contract

* deployed new factory

* committing progress

* updating arbiter to be CC safe

* merge branches

---------

+++ contracts/interfaces/IOracle.sol
@@ -2,8 +2,8 @@
 
 interface IOracle {
     /** current price for token asset. denominated in USD */
-    function getLatestAnswer(address token) external returns (int);
+    function getLatestAnswer(address token) external returns (int256);
 
     /** Readonly function providing the current price for token asset. denominated in USD */
-    function _getLatestAnswer(address token) external view returns (int);
+    function _getLatestAnswer(address token) external view returns (int256);
 }
